+++ src/app/app.config.ts
@@ -4,18 +4,7 @@ import {
   provideAppInitializer,
   provideBrowserGlobalErrorListeners,
 } from '@angular/core';
-import { MaintenanceService } from '@/app/maintenance';
-
-export function initAppFactory(ms: MaintenanceService) {
-  return () => ms.init();
-}
 
 export const appConfig: ApplicationConfig = {
-  providers: [
-    provideBrowserGlobalErrorListeners(),
-    provideAppInitializer(() => {
-      const ms = inject(MaintenanceService);
-      return ms.init();
-    }),
-  ],
+  providers: [provideBrowserGlobalErrorListeners()],
 };
