+++ src/app/app.html
@@ -1,5 +1,5 @@
-@if(ms.enabled){<maintenance></maintenance>}
 <main id="root">
+  <maintenance></maintenance>
   <ngx-sonner-toaster
     position="top-right"
     [toastOptions]="{
