+++ src/app/app.ts
@@ -1,11 +1,10 @@
-import { Component, signal } from '@angular/core';
+import { Component } from '@angular/core';
 
 import { NgxSonnerToaster } from 'ngx-sonner';
 
 import { Flask } from '@/app/flask/flask';
 import { Dropzone } from '@/app/dropzone/dropzone';
 import { Maintenance } from '@/app/maintenance/maintenance';
-import { MaintenanceService } from './maintenance';
 
 @Component({
   selector: 'app',
@@ -13,8 +12,4 @@ import { MaintenanceService } from './maintenance';
   styleUrl: 'app.css',
   imports: [Flask, Dropzone, NgxSonnerToaster, Maintenance],
 })
-export class App {
-  protected readonly title = signal('Bundler');
-
-  constructor(public ms: MaintenanceService) {}
-}
+export class App {}
