+++ src/app/maintenance/maintenance.html
@@ -1,3 +1,4 @@
+@if (isEnabled()) {
 <div class="maintenance-container">
   <div class="maintenance-box">
     <img class="maintenance-logo" src="logo.svg" />
@@ -15,3 +16,4 @@
     >
   </div>
 </div>
+}
