+++ src/app/maintenance/maintenance.ts
@@ -1,9 +1,14 @@
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
+import { MaintenanceService } from '@/services/maintenance';
 
 @Component({
   selector: 'maintenance',
   imports: [],
   templateUrl: 'maintenance.html',
   styleUrl: 'maintenance.css',
+  standalone: true,
 })
-export class Maintenance {}
+export class Maintenance {
+  ms = inject(MaintenanceService);
+  isEnabled = this.ms.enabled;
+}
